Running `pipenv` in parallel significantly improved the performance of `pipenv` source, on my 8 core (4 cores + ht) machine running `./script/test pipenv` takes:

2.5.0:
 - 25.979721s
 - 25.930273s
 - 26.492875s

2.5.0 + parallel (threads):
 - 8.658091s
 - 8.751204s
 - 8.751976s