@krzysztof-pawlik-gat 2019-10-23
> sparkles this is great, thank you! So great, in fact, that I'm going to extend it as a pattern to the other sources and possibly the parent `Source` class.

Sounds good, I think that `pip` source could use this too (from a quick glance), `npm` probably not as it invokes the binary once anyway. I didn't look at other sources, to be honest.

> I'm going pushing a new release and will merge this PR after. Due to CI not running on forks, I may ping you in an issue if this ends up with CI problems after merging to master.

I was surprised by that too! If there are any issues with this change I'll be happy to fix those.